formula_support: split test KegOnlyReasonTests into test os_mac_formula_support
